validaciones emergentes

+++ Entity/Model/Security/Person.cs
@@ -8,7 +8,7 @@ namespace Entity.Model.Security
     public class Person : ABaseEntity
     {
         public int DocumentTypeId { get; set; }
-        public int Identification { get; set; }
+        public long Identification { get; set; }
         public string FisrtName { get; set; }
         public string SecondName { get; set; }
         public string LastName { get; set; }
